+++ packages/angular/src/canvas/qr.component.ts
@@ -269,6 +269,7 @@ export class Qr implements OnInit, OnChanges {
 
   ngOnInit() {
     this.ctx = this.canvas.nativeElement.getContext("2d")!;
+    this.generateQr();
   }
 
   ngOnChanges() {