+++ packages/angular/src/public-api.ts
@@ -2,5 +2,5 @@
  * Public API Surface of angular
  */
 
-export { Qr as CanvasQr } from "./canvas/qr.component";
-export { Qr as SvgQr } from "./svg/qr.component";
+export { Qr as CanvasQr } from "./canvas";
+export { Qr as SvgQr } from "./svg";