+++ packages/core/scripts/modify-package.mjs
@@ -4,5 +4,9 @@ const pkg = JSON.parse(await fs.readFile("./package.json"));
 const pkgProd = JSON.parse(await fs.readFile("./package-prod.config"));
 
 const result = { ...pkg, ...pkgProd };
+delete result.devDependencies;
+delete result.scripts;
+
+await fs.writeFile("./package.json", JSON.stringify(result, null, 2));
 
 await fs.writeFile("./package.json", JSON.stringify(result, null, 2));