+++ packages/angular/CHANGELOG.md
@@ -5,3 +5,9 @@
 ### Major Changes
 
 - Beta release
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Fix Canvas qr