+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qrgrid/cli
 
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- CI/CD remove devDependency and scripts from package during build
+- Updated dependencies
+  - @qrgrid/core@0.2.2
+
 ## 1.0.0-beta
 
 ### Major Changes