+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qrgrid/react
 
+## 1.2.1
+
+### Patch Changes
+
+- CI/CD remove devDependency and scripts from package during build
+- Updated dependencies
+  - @qrgrid/core@0.2.2
+
 ## 1.2.0
 
 ### Minor Changes