+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qrgrid/vue
 
+## 1.3.1
+
+### Patch Changes
+
+- CI/CD remove devDependency and scripts from package during build
+- Updated dependencies
+  - @qrgrid/core@0.2.2
+
 ## 1.3.0
 
 ### Minor Changes